chore(core): move index.d.ts to src
otherwise eslint import plugin failed when extending type
from core
